Fix to update the Mobile data in settings db correctly.
With the current code mobile data is getting automatically
enabled when disabled and rebooted.This is happening due
to the wrong updation of mobile data value in settings db.
In DcTracker, due to the delay in handling the message the
initial dummy sub id data value is wrongly getting updated
to the valid sub id.

Fix: In PhoneBase, allow get and set data enabled operations
only if sub id is valid and not a dummy sub id.

Change-Id: I0f53f73f0951ea73f85ce8056a204c809c6bec1c
CRs-Fixed: 945072.